requirements.txt
requests>=2.28.0
python-dotenv>=1.0.0
colorama>=0.4.6
hyperliquid-python-sdk>=0.7.0
eth-account>=0.10.0
eth-utils>=2.0.0
websocket-client>=1.6.0
anthropic>=0.40.0
